requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
gunicorn
pymongo
Werkzeug
